+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/GradleProperties.kt
@@ -3,22 +3,22 @@
 package org.jetbrains.intellij.platform.gradle
 
 import org.gradle.api.provider.ProviderFactory
+import org.jetbrains.intellij.platform.gradle.Constants.Locations
 import org.jetbrains.intellij.platform.gradle.Constants.Plugin
 import org.jetbrains.intellij.platform.gradle.extensions.IntelliJPlatformExtension
 import org.jetbrains.intellij.platform.gradle.tasks.BuildSearchableOptionsTask
 import org.jetbrains.intellij.platform.gradle.utils.Logger
 
 /**
- * The IntelliJ Platform Gradle Plugin build features dedicated to control some of the low-level Gradle plugin behaviors.
- * To enable or disable a particular feature, add a Project property to the `gradle.properties` file with the following pattern:
+ * The IntelliJ Platform Gradle Plugin introduces custom Gradle properties to control some of the low-level Gradle plugin behaviors.
+ * To adjust a particular feature, add a Project property to the `gradle.properties` file with the following pattern:
  *
  * ```
- * org.jetbrains.intellij.platform.buildFeature.<buildFeatureName>=<true|false>
+ * org.jetbrains.intellij.platform.<propertyName>=<value>
  * ```
- *
- * Switch to [org.gradle.api.configuration.BuildFeatures] when supporting Gradle 8.5+.
  */
-enum class BuildFeature(private val defaultValue: Boolean) {
+sealed class GradleProperties<T : Any>(val defaultValue: T) {
+
     /**
      * Instruct the IDE that sources are needed to be downloaded when working with IntelliJ Platform Gradle Plugin.
      * Value is passed directly to the [Idea Gradle Plugin](https://docs.gradle.org/current/userguide/idea_plugin.html)
@@ -26,20 +26,28 @@ enum class BuildFeature(private val defaultValue: Boolean) {
      *
      * @see <a href="https://docs.gradle.org/current/dsl/org.gradle.plugins.ide.idea.model.IdeaModule.html#org.gradle.plugins.ide.idea.model.IdeaModule:downloadSources">IdeaModule.downloadSources</a>
      */
-    DOWNLOAD_SOURCES(true),
+    object DownloadSources : GradleProperties<Boolean>(true)
+
+    object IntellijPlatformCache : GradleProperties<String>("")
+
+    object LocalPlatformArtifacts : GradleProperties<String>("")
 
     /**
-     * When the [BuildSearchableOptionsTask] doesn't produce any results, e.g., when the plugin doesn't implement any settings, a warning is shown to suggest disabling it
-     * for better performance with [IntelliJPlatformExtension.buildSearchableOptions].
+     * When the [BuildSearchableOptionsTask] doesn't produce any results, e.g., when the plugin doesn't implement any settings, a warning is shown
+     * to suggest disabling it for better performance with [IntelliJPlatformExtension.buildSearchableOptions].
      */
-    NO_SEARCHABLE_OPTIONS_WARNING(true),
+    object NoSearchableOptionsWarning : GradleProperties<Boolean>(true)
 
     /**
      * Due to IDE limitations, it is impossible to run the IDE in headless mode to collect searchable options for a paid plugin.
      * As paid plugins require providing a valid license and presenting a UI dialog, it is impossible to handle such a case, and the task will fail.
      * This feature flag displays the given warning when the task is run by a paid plugin.
      */
-    PAID_PLUGIN_SEARCHABLE_OPTIONS_WARNING(true),
+    object PaidPluginSearchableOptionsWarning : GradleProperties<Boolean>(true)
+
+    object ProductsReleasesAndroidStudioUrl : GradleProperties<String>(Locations.PRODUCTS_RELEASES_ANDROID_STUDIO)
+
+    object ProductsReleasesJetBrainsIdesUrl : GradleProperties<String>(Locations.PRODUCTS_RELEASES_JETBRAINS_IDES)
 
     /**
      * Checks whether the currently used Gradle IntelliJ Plugin is outdated and if a new release is available.
@@ -49,9 +57,10 @@ enum class BuildFeature(private val defaultValue: Boolean) {
      * If the current version is outdated, the plugin will emit a warning with its current and the latest version.
      *
      * Feature respects the Gradle `--offline` mode.
-     *
      */
-    SELF_UPDATE_CHECK(true),
+    object SelfUpdateCheck : GradleProperties<Boolean>(true)
+
+    object ShimServerPort : GradleProperties<Int>(7348)
 
     /**
      * By default, JetBrains Cache Redirector is used when resolving Maven repositories or any resources used by the IntelliJ Platform Gradle Plugin.
@@ -59,36 +68,24 @@ enum class BuildFeature(private val defaultValue: Boolean) {
      *
      * It is possible to refer to the direct location (whenever it is possible) by switching off JetBrains Cache Redirector globally.
      */
-    USE_CACHE_REDIRECTOR(true);
+    object UseCacheRedirector : GradleProperties<Boolean>(true)
 
-    fun getValue(providers: ProviderFactory) = providers.gradleProperty(toString())
-        .map { it.toBoolean() }
-        .orElse(defaultValue)
-
-    override fun toString() = name
-        .lowercase()
-        .split('_')
-        .joinToString(
-            separator = "",
-            transform = { it.replaceFirstChar { c -> c.uppercase() } },
-        )
-        .replaceFirstChar { c -> c.lowercase() }
-        .let { "${Plugin.ID}.buildFeature.$it" }
-
-    /**
-     * Checks if the specified build feature is enabled for the current project.
-     *
-     * @param providers Gradle [ProviderFactory] instance.
-     * @return A provider containing the boolean value
-     */
-    fun isEnabled(providers: ProviderFactory) =
-        getValue(providers).map { value ->
-            value.also {
-                val log = Logger(BuildFeature::class.java)
-                when (value) {
-                    true -> log.info("Build feature is enabled: $this")
-                    false -> log.info("Build feature is disabled: $this")
-                }
-            }
-        }
+    override fun toString() =
+        requireNotNull(this::class.simpleName)
+            .replaceFirstChar(Char::lowercase)
+            .let { Plugin.ID + '.' + it }
 }
+
+inline operator fun <reified T : Any> ProviderFactory.get(property: GradleProperties<T>) =
+    gradleProperty(property.toString())
+        .map {
+            Logger(javaClass).info("Read Gradle property: $property=$it")
+
+            when (property.defaultValue) {
+                is Boolean -> it.toBoolean()
+                is Int -> it.toInt()
+                is String -> it
+                else -> it
+            } as T
+        }
+        .orElse(property.defaultValue)